% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{add index column}
\usage{
add_index(x, val, index_year, log_level, log_appender)
}
\arguments{
\item{x}{Year_sector_data_object}

\item{val}{Column in x$df containing values to be indexed}

\item{index_year}{Year to start indexing. uses min x$years by default}

\item{log_level}{The severity level at which log messages are written from
least to most serious: TRACE, DEBUG, INFO, WARN, ERROR, FATAL. Default is
level is INFO. See \code{?flog.threshold()} for additional details.}

\item{log_appender}{Defaults to write the log to "console", alternatively you
can provide a character string to specify a filename to also write to. See
for additional details \code{?futile.logger::appender.file()}.}
}
\description{
NOTE: THIS FUNCTION RELIES ON DATA WHICH ARE CLASSIFIED AS
  OFFICIAL-SENSITIVE. THE OUTPUT OF THIS FUNCTION IS AGGREGATED, AND
  PUBLICALLY AVAILABLE IN THE FINAL STATISTICAL RELEASE, HOWEVER CARE MUST BE
  EXERCISED WHEN CREATING A PIPELINE INCLUDING THIS FUNCTION. IT IS HIGHLY
  ADVISEABLE TO ENSURE THAT THE DATA WHICH ARE CREATED BY THE \code{extract_}
  FUNCTIONS ARE NOT STORED IN A FOLDER WHICH IS A GITHUB REPOSITORY TO
  MITIGATE AGAINST ACCIDENTAL COMMITTING OF OFFICIAL DATA TO GITHUB. TOOLS TO
  FURTHER HELP MITIGATE THIS RISK ARE AVAILABLE AT
  https://github.com/ukgovdatascience/dotfiles.
}
\details{
Uses \code{val} and year columns to produce new column,
containing an index of \code{val} starting at the year specified by
\code{index_year}. The column is named index, suffixed by the name of
\code{val}. Returns x$df with additional column.
}
