% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_timeseries.R
\name{APS_timeseries}
\alias{APS_timeseries}
\title{APS time series maker}
\usage{
APS_timeseries(...)
}
\arguments{
\item{...}{either the locations fo the APS SPSS files from which you want
to form the time series, or the data frames you want to form into the time
series list.}
}
\value{
The function a list of APS data frames, with each list element named
  after the year from which the data come
}
\description{
Builds a list of APS data frames by year that can be passed to
  other table making functions

  IT IS HIGHLY ADVISEABLE TO ENSURE THAT THE DATA WHICH ARE CREATED BY THIS
  FUNCTION ARE NOT STORED IN A FOLDER WHICH IS A GITHUB REPOSITORY TO
  MITIGATE AGAINST ACCIDENTAL COMMITTING OF OFFICIAL DATA TO GITHUB. TOOLS TO
  FURTHER HELP MITIGATE THIS RISK ARE AVAILABLE AT
  https://github.com/ukgovdatascience/dotfiles.
}
\details{
The best way to understand what happens when you run this function
  is to look at the source code, which is available at
  \url{https://github.com/ukgovdatascience/eesectors/blob/master/R/}. The
  code is relatively transparent and well documented. A brief explanation of
  what the function does here:

  1. The function checks if the inputs are file directories or data frames

  2. The function finds the year of each inputted dataset

  3. The function returns a list of datasets named per year (x$y20XX)
}
\examples{

\dontrun{
library(eesectors)
extract_APS_data(...)
}


}
