% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bounds.R
\name{calculate_bounds}
\alias{calculate_bounds}
\title{Helper function to calculate upper and lower bounds}
\usage{
calculate_bounds(df, measure = NULL, test_year = NULL, tol = 3)
}
\arguments{
\item{df}{Dataframe to test.}

\item{measure}{Quoted variable name of interest from \code{df}.}

\item{test_year}{Year of interest to test.}

\item{tol}{Round output to this many digits.}
}
\value{
Returns the percentage difference of \code{y} relative to \code{x}.
}
\description{
Calculates the upper bounds of acceptable values based on the
  \code{median(x) + mad(x) * tol} where \code{is.integer(tol) == TRUE} and
  the obverse lower bound, and presents the upper and lower bounds in a
  dataframe with the variable of interest.
}
\details{
Provides the basis of tests for the most recently calculated values
  of Gross Value Added (GVA), exports, etc. The rationale is that this year's
  statistics should be fairly similar to last year's, and we can use
  statistical tests to identify possibly erroneous values.

  Statistical tests can be as complicated as we desire, but in the present
  version the test is a simple outlier detection based on identifying values
  that are more than \code{x} times the median absolute deviation
  (\code{mad()}), where \code{x} is defined by the \code{tol()} argument.
  This defaults to three, but can be adjusted for more or less stringent
  outlier tests.

  Summary statistics like \code{mean} and \code{mad} are computed on the data
  prior to \code{test_year}, equivalent to: \code{df[df$year < test_year,]}.
}
\references{
Leys et al. (2013), 'Detecting outliers: Do not use standard
  deviation around the mean, use absolute deviation around the median',
  Journal of Experimental Social Psychology. 49 (4), pp. 764 - 766.
  http://dx.doi.org/10.1016/j.jesp.2013.03.013
}

