% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_sic}
\alias{clean_sic}
\title{Clean SIC codes}
\usage{
clean_sic(x)
}
\arguments{
\item{x}{A character vector of SIC codes.}
}
\value{
A cleaned character vector of SIC codes.
}
\description{
Converts 3 or 4 digit SIC codes from format \code{123} or
  \code{1234} into \code{12.3} and \code{12.34} respectively. Codes with
  length of 2 or greater than 4 are ignored, and returned as is.
}
