% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_GVA_long.R
\name{combine_GVA_long}
\alias{combine_GVA_long}
\title{Combine GVA, ABS, SIC91 datasets in long format -
  retaining separate rows for each SIC code}
\usage{
combine_GVA_long(ABS = NULL, GVA = NULL, SIC91 = NULL,
  DCMS_sectors = eesectors::DCMS_sectors)
}
\arguments{
\item{ABS}{ABS data as extracted by \code{eesectors::extract_ABS_data()}.}

\item{GVA}{ABS data as extracted by \code{eesectors::extract_GVA_data()}.}

\item{SIC91}{ABS data as extracted by \code{eesectors::extract_SIC91_data()}.}

\item{DCMS_sectors}{ABS data as extracted by
\code{eesectors::extract_DCMS_sectors()} or matching the
\code{eesectors::DCMS_sectors} in-built dataset.}
}
\description{
Combines datasets exracted from the underlying spreadsheet using
  the \code{extract_XXX} functions.

  NOTE: THIS FUNCTION RELIES ON DATA WHICH ARE CLASSIFIED AS
  OFFICIAL-SENSITIVE. THE OUTPUT OF THIS FUNCTION IS AGGREGATED, AND
  PUBLICALLY AVAILABLE IN THE FINAL STATISTICAL RELEASE, HOWEVER CARE MUST BE
  EXERCISED WHEN CREATING A PIPELINE INCLUDING THIS FUNCTION. IT IS HIGHLY
  ADVISEABLE TO ENSURE THAT THE DATA WHICH ARE CREATED BY THE \code{extract_}
  FUNCTIONS ARE NOT STORED IN A FOLDER WHICH IS A GITHUB REPOSITORY TO
  MITIGATE AGAINST ACCIDENTAL COMMITTING OF OFFICIAL DATA TO GITHUB. TOOLS TO
  FURTHER HELP MITIGATE THIS RISK ARE AVAILABLE AT
  https://github.com/ukgovdatascience/dotfiles.
}
\details{
The best way to understand what happens when you run this function
  is to view the source code.
}
