% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_working_file}
\alias{example_working_file}
\title{Get path to example_working_file.xlsx example}
\usage{
example_working_file(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
eesectors comes bundled with an example file in its `inst/extdata`
directory. This function returns its path.
}
\examples{
example_working_file()
example_working_file("example_working_file.xlsx")
}
