% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports_by_sector_table.R
\name{exports_by_sector_table}
\alias{exports_by_sector_table}
\title{Format exports by sector table}
\usage{
exports_by_sector_table(x, max_year = 2014, html = TRUE)
}
\arguments{
\item{x}{Input dataframe.}

\item{max_year}{The maximum year for which measurements are available for the
given measure. This is experimental, and likely to be removed in favour of
the function determining this for itself from teh data.}

\item{html}{Should the table be output as is or using \code{xtable} as html.}
}
\value{
Properly formatted exports table for release.
}
\description{
\code{exports_by_sector_table} Formats the exports by sectors
  table labelled as 3.1 in the 2016 Economic estimates of DCMS Sectors
  statistical release
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf}).
}
\details{
\code{exports_by_sector_table} takes as input a standardised long
  format exports data frame, and uses \code{xtable} to format a pretty html
  table.
}
\examples{

library(dplyr)
library(eesectors)

exports_by_sector_table(exports_by_sector_2016)

}

