% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ABS_data.R
\name{extract_APS_data}
\alias{extract_APS_data}
\title{Extract APS data from ONS Annual Population Survey SPSS .sav file}
\usage{
extract_APS_data(x)
}
\arguments{
\item{x}{Location of the input sav file. Named something like
"APSP_JDXX_CLIENT_PWTA14.sav".}
}
\value{
The function returns a dataframe of the variables used for the Economic Estimates
employment statistics.
}
\description{
The APS data which underlies the Economic Sectors employemnt statistics for DCMS sectors
  data is provided to DCMS as a SPSS .sav file from the Office for
  National Statistics. This function extracts the APS data from that
  SPSS file and return the necessary data as a dataframe.

  You can obtain APS data from the UK Data Service here
  \url{https://discover.ukdataservice.ac.uk/series/?sn=200002}, where guide to variables
  by year can also be found. If you want change the variables this code extracts from
  the APS data, start witht these guides to find what the variables are called.

  IT IS HIGHLY ADVISEABLE TO ENSURE THAT THE DATA WHICH ARE CREATED BY THIS
  FUNCTION ARE NOT STORED IN A FOLDER WHICH IS A GITHUB REPOSITORY TO
  MITIGATE AGAINST ACCIDENTAL COMMITTING OF OFFICIAL DATA TO GITHUB. TOOLS TO
  FURTHER HELP MITIGATE THIS RISK ARE AVAILABLE AT
  https://github.com/ukgovdatascience/dotfiles.
}
\details{
The best way to understand what happens when you run this function
  is to look at the source code, which is available at
  \url{https://github.com/ukgovdatascience/eesectors/blob/master/R/}. The
  code is relatively transparent and well documented. A brief explanation of
  what the function does here:

  1. The function calls \code{foreign::read.spss} to load the data from the
  source file.

  2. An array of the names of variables used for the analysis is created. These
  should be:

  3. The columns of data named in the array are extracted from the loaded SPSS data
  resulting in a dataframe of only the necessary data.

  4. The dataframe created is returned.
}
\examples{

\dontrun{
library(eesectors)
APS_data_2015 <- extract_ABS_data("APSP_JD15_CLIENT_PWTA14.sav")
}

}
