% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3.2.R
\name{figure3.2}
\alias{figure3.2}
\title{Create Figure 3.2}
\usage{
figure3.2(x, ...)
}
\arguments{
\item{x}{Object of class \code{long_table()}.}

\item{...}{Passes arguments to \code{ggplot}.}
}
\value{
Figure 3.2
}
\description{
Creates Figure 3.2 from
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf})
  using \code{ggplot2}.
}
\details{
\code{figure3.2} takes as input a standardised long format GVA data
  frame, and uses \code{ggplot} and \code{govstyle} to format a pretty(ish) plot.
}
\examples{

library(eesectors)

gva <- year_sector_data(GVA_by_sector_2016)
figure3.2(gva)

}
