% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{format_table()}
\usage{
format_table(x)
}
\arguments{
\item{x}{Object of \code{class(x) == 'long_data'}.}
}
\value{
wide format table
}
\description{
Method to convert table into wide format
}
\details{
Method to convert table into wide format
}

