% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GVA_notebook.R
\name{get_GVA_notebook}
\alias{get_GVA_notebook}
\title{Downloads the latest version of the combine_GVA.Rmd notebook}
\usage{
get_GVA_notebook(url = NULL, file = "combine_GVA.Rmd")
}
\arguments{
\item{url}{An alternate url for the \code{combine_GVA.Rmd} notebook.}

\item{file}{Local file where the downloaded \code{combine_GVA.Rmd} will be
stored.}
}
\value{
The function returns nothing, but saves output to the location
  specified by the \code{file} argument.
}
\description{
Downloads the latest version of the combine_GVA.Rmd notebook,
  which is stored within the package at \code{inst/combine_GVA.Rmd}.
}
\details{
Since combining the various datasets is quite a complex operation
  that could easily be disrupted by small changes to the input spreadhseet,
  it may be easier to run the combine operation within an R notebook. This
  function uses the \code{httr} package to download the latest version of the
  notebook from the master branch of the
  \url{https://github.com/ukgovdatascience/eesectors} repository.
}
\examples{

\dontrun{
library(eesectors)
get_GVA_notebook()
}

}
