% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA.R
\docType{data}
\name{GVA}
\alias{GVA}
\title{GVA data for 2010:2015.}
\format{A tibble with 54 rows and 3 variables: \describe{
  \item{sector}{DCMS sector}
  \item{year}{calendar year}
  \item{GVA}{gross value added}
  }.}
\source{
\url{https://www.gov.uk/government/statistics/dcms-sectors-economic-estimates-2016}
}
\usage{
GVA
}
\description{
Data extracted from the 2016 Economic Estimates of DCMS Sectors 2016 report.
These data are provided for testing ouputs of this package against, and for
the initial development phase during which the original data sources were not
available.
}
\keyword{DCMS}
\keyword{Economic}
\keyword{Estimates}
\keyword{Sectors}

