% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA_by_sector.R
\name{GVA_by_sector}
\alias{GVA_by_sector}
\title{Calculate GVA by sector}
\usage{
GVA_by_sector(combine_GVA_long = NULL, GVA = NULL, tourism = NULL)
}
\arguments{
\item{combine_GVA_long}{data output from \code{eesectors::combine_GVA_long()}.}

\item{GVA}{ABS data as extracted by \code{eesectors::extract_GVA_data()}.}

\item{tourism}{ABS data as extracted by \code{eesectors::extract_tourism_data()}.}
}
\description{
Combines datasets exracted from the underlying spreadsheet using
  the \code{extract_XXX} functions.

  NOTE: THIS FUNCTION RELIES ON DATA WHICH ARE CLASSIFIED AS
  OFFICIAL-SENSITIVE. THE OUTPUT OF THIS FUNCTION IS AGGREGATED, AND
  PUBLICALLY AVAILABLE IN THE FINAL STATISTICAL RELEASE, HOWEVER CARE MUST BE
  EXERCISED WHEN CREATING A PIPELINE INCLUDING THIS FUNCTION. IT IS HIGHLY
  ADVISEABLE TO ENSURE THAT THE DATA WHICH ARE CREATED BY THE \code{extract_}
  FUNCTIONS ARE NOT STORED IN A FOLDER WHICH IS A GITHUB REPOSITORY TO
  MITIGATE AGAINST ACCIDENTAL COMMITTING OF OFFICIAL DATA TO GITHUB. TOOLS TO
  FURTHER HELP MITIGATE THIS RISK ARE AVAILABLE AT
  https://github.com/ukgovdatascience/dotfiles.
}
\details{
The best way to understand what happens when you run this function
  is to view the source code.
}
