% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GVA_by_sector_2016}
\alias{GVA_by_sector_2016}
\title{GVA data for 2010:2015.}
\format{A tibble with 54 rows and 3 variables:
\describe{
  \item{sector}{DCMS sector, one of: \code{c('creative','culture','digital',
  'gambling','sport','telecoms','tourism', 'UK', 'all_dcms')}, note that
  \code{all_dcms} is not the sum of the other sectors, due to overlap between
   sectors.}
  \item{year}{calendar year.}
  \item{GVA}{gross value added (GBP millions).}
}}
\source{
\url{https://www.gov.uk/government/statistics/dcms-sectors-economic-estimates-2016}
}
\usage{
GVA_by_sector_2016
}
\description{
Data extracted from the 2016 Economic Estimates of DCMS Sectors 2016 report.
These data are provided for testing ouputs of this package against, and for
the initial development phase during which the original data sources were not
available.
}
\keyword{DCMS}
\keyword{Economic}
\keyword{Estimates}
\keyword{Sectors}
