% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA_by_sector_table.R
\name{GVA_by_sector_table}
\alias{GVA_by_sector_table}
\title{Format GVA by sectors table}
\usage{
GVA_by_sector_table(x, html = TRUE)
}
\arguments{
\item{x}{Input dataframe.}

\item{html}{Should the table be output as is or using \code{xtable} as html.}
}
\value{
Properly formatted GVA table for release.
}
\description{
\code{GVA_by_sector_table} Formats the GVA by sectors table labelled as
  3.1 in the 2016 Economic estimates of DCMS Sectors statistical release
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf}).
}
\details{
\code{GVA_by_sector_table} takes as input a standardised long format GVA data frame, and uses
  \code{xtable} to format a pretty html table.
}
\examples{

library(dplyr)
library(eesectors)

GVA_by_sector_table(GVA_by_sector_2016)

}

