% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA_plot_2.R
\name{GVA_plot_2}
\alias{GVA_plot_2}
\title{Create GVA plot 2}
\usage{
GVA_plot_2(x)
}
\arguments{
\item{x}{Input dataframe.}
}
\value{
Properly formatted GVA table for release.
}
\description{
\code{GVA_plot_2} Creates Figure 3.2 from
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf})
  using \code{ggplot2}.
}
\details{
\code{GVA_plot_2} takes as input a standardised long format GVA data
  frame, and uses \code{ggplot} and \code{govstyle} to format a pretty(ish) plot.
}
\examples{

library(EESectors)

GVA_plot_2(GVA_by_sector_2016)

}

