% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA_table.R
\docType{data}
\name{GVA_table}
\alias{GVA_table}
\title{GVA by sectors (Table 3.1).}
\format{A tibble with 10 rows and 10 variables.}
\source{
\url{https://www.gov.uk/government/statistics/dcms-sectors-economic-estimates-2016}
}
\usage{
GVA_table
}
\description{
Table 3.1 from the 2016 Economic Estimates for DCMS Sectors publication. This
dataset is included for testing purposes.
}
\keyword{DCMS}
\keyword{Economic}
\keyword{Estimates}
\keyword{Sectors}

