% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrity_check}
\alias{integrity_check}
\title{Check success of type conversion: character to numeric}
\usage{
integrity_check(x)
}
\arguments{
\item{x}{vector of \code{length(x) > 1} and of nominally of
\code{is.character(x) == TRUE}.}
}
\value{
Returns the \code{conversion_issues} vector, with following
  attribites: \code{length(x) == length(conversion_issues)} and
  \code{is.logical(conversion_issues) == TRUE}.
}
\description{
Checks which values will convert to \code{NA} when converting a
  vector from \code{character} to \code{numeric} or \code{integer}.
}
\details{
The spreadsheets which come from the Office for National Statistics
  (ONS) often contain characters within what should be purely numeric fields.
  When we convert these to numeric these will be converted to \code{NA}, but
  it is not entirely clear after the fact which values have been converted or
  why. This function checks in advance which values will be converted to
  \code{NA} so we can confirm that it is nothing to worry about.

  IT WILL ONLY RAISE A WARNING IF: there are new NAs, and any of these NAs,
  can cleanly be converted to numeric (wihtout turning to \code{NA}) - i.e.
  they are actually numbers, and not characters, which is what we would
  expect.
}
