% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_tables.R
\name{jobs_timeseries}
\alias{jobs_timeseries}
\title{Sector employment time series table}
\usage{
jobs_timeseries(..., tourism = NULL, sectors = eesectors::DCMS_sectors)
}
\arguments{
\item{...}{APS time series `data.frames`s or the file destinations of APS datasets
(providing APS data sets will cause the function to run slowly as it loads the data)}
}
\value{
The function returns the sector employment time series table
}
\description{
Produces a time series table of employment by DCMS sector
}
\details{
The function \code{eesectors::jobs_timeseries} produces a basic time series
  of data based on jobs per DCMS sector.
}
\examples{

\dontrun{
library(eesectors)
employment_timeseries <- jos_timeseries(
aps2011, aps2012, aps2013, aps2014, aps2015
)
}

\dontrun{
library(eesectors)
employment_timeseries <- jos_timeseries(
"./APS_data_2013.sav","./APS_data_2014.sav","./APS_data_2015.sav",
)
}

}
