% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_data.R
\name{long_data}
\alias{long_data}
\title{long data class}
\usage{
long_data(x)
}
\arguments{
\item{x}{Input dataframe, see details.}
}
\value{
If the class is instantiated correctly, nothing is returned.
}
\description{
\code{long_data} is the class used for the creation of tables
  3.1, 4.1, 4.5, and 5.1 of the DCMS Sectors Economic Estimate
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf}).
}
\details{
The \code{long_data} class expects a \code{data.frame} with three
  columns: sector, year, and measure, where measure is one of GVA, exports,
  or enterprises. The \code{data.frame} should include historical data, which
  is used for checks on the quality of this year\'s data, and for producing
  tables and plots.

  Once inititated, the class has five slots: \code{df}: the basic
  \code{data.frame}, \code{colnames}: a character vector containing the
  column names from the \code{df}, \code{type}: a character vector of
  \code{length(type) == 1} describing the type of data (\"GVA\", \"exports\",
  or \"enterprises\") imputed from column names, \code{sector_levels}: a
  factor vector containing levels of \code{df$sector} of the factor sector,
  \code{years}: an integer vector containing \code{unique(df$year)}.
}
\examples{

library(EESectors)

GVA <- long_data(GVA_by_sector_2016)

}

