% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maha_check.R
\name{maha_check}
\alias{maha_check}
\title{maha check}
\usage{
maha_check(x)
}
\arguments{
\item{x}{Input dataframe, see details.}
}
\value{
If the class is instantiated correctly, nothing is returned.
}
\description{
\code{maha_check} is a function that runs the mahalanobis check
}
\details{
this is based on \code{assertr::insist_rows} and implements a
  malahanobis distance check using \code{assertr:maha_dist}.
}
