% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_utils.R
\name{main_job_mask}
\alias{main_job_mask}
\title{Economic activity filters}
\usage{
main_job_mask(x)
}
\arguments{
\item{x}{a `data.frame` as extracted by \code{eesectors::extract_APS_data}}
}
\value{
The function returns the APS `data.frame` with only those
  with a first or second job depending on functions used.
}
\description{
Filters out people based on their main/second job
  economic activity status
}
\details{
Filters out people based on their main/second job
  economic activity status - INECAC05 for first job and SECJMBR for second
  job (1 = full time, 2 = part time).
}
\examples{

\dontrun{
library(eesectors)
APS_data_jobs=sector_jobs(APS_data)
)
}

}
