% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_cols}
\alias{na_cols}
\title{Find NAs in dataframe columns}
\usage{
na_cols(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
Character vector containing the names of columns which contain \code{NA}.
}
\description{
Looks through a \code{data.frame} and returns a character vector
  of column names corresponding to columns which contain \code{NA}.
}

