% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_utils.R
\name{regionsRecode}
\alias{regionsRecode}
\title{Recode functions}
\usage{
regionsRecode(x)
}
\arguments{
\item{x}{a `data.frame` column as extracted by \code{eesectors::extract_APS_data}}
}
\value{
The function returns the APS variable as a dataframe column with recoded
regional variables
}
\description{
Recodes various APS functions into ones usable
for the analysis
}
\details{
These functions recode APS variables into variales
useful for the statistical release, for example, the regional variables
into NUTS1/GOR regions
}
\examples{

\dontrun{
library(eesectors)
APS_data_jobs$var<-regionsRecode(APS_data$var)
)
}

}
