% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_to.R
\name{relative_to}
\alias{relative_to}
\title{Helper function to calculate relative values}
\usage{
relative_to(x, y, digits = 2)
}
\arguments{
\item{x}{base variable.}

\item{y}{relative variable.}

\item{digits}{Round output to this many digits.}
}
\value{
Returns the percentage difference of \code{y} relative to \code{x}.
}
\description{
\code{relative_to} Calculates the percentage difference in
  \code{y}, relative to \code{x}
}
\details{
\code{relative_to} Calculates the percentage difference in
  \code{y}, relative to \code{x}
}

