% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundf}
\alias{roundf}
\title{Prepare values for table}
\usage{
roundf(x, fmt = "\%.1f")
}
\arguments{
\item{x}{Value to be transformed.}

\item{fmt}{Inherits from \code{sprintf()}: the format to be returned.}
}
\value{
Tranfsormed variable \code{x}.
}
\description{
Prepares values for publication in wide tables by dividing by
  1000 and rounding and truncating using \code{sprintf()}. Vectorised to work
  on single values, vectors, and data.frames.
}
\seealso{
\code{\link{sprintf}}.
}

