% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_rds}
\alias{save_rds}
\title{Save and check Rds files}
\usage{
save_rds(x, full_path)
}
\arguments{
\item{x}{Input dataframe.}

\item{full_path}{Path to which Rds file will be saved.}
}
\value{
Returns nothing
}
\description{
Saves a dataframe \code{x} to an Rds file, checks that the file
  was produced, and raises a warning if it was not.
}
