% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_utils.R
\name{sector_jobs}
\alias{sector_jobs}
\title{Sector jobs}
\usage{
sector_jobs(x)
}
\arguments{
\item{x}{a `data.frame` as extracted by \code{eesectors::extract_APS_data}}
}
\value{
The function returns the APS data as a dataframe with the sector
  job count variables appended
}
\description{
Counts the number of jobs a person has in each sector (0, 1 or 2)
}
\details{
This function sums up the jobs per sector found using the appendSectors
  function.
}
\examples{

\dontrun{
library(eesectors)
APS_data_jobs=sector_jobs(APS_data)
)
}

}
