% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APS_utils.R
\name{sector_jobs_count}
\alias{sector_jobs_count}
\title{Sector jobs count}
\usage{
sector_jobs_count(x)
}
\arguments{
\item{x}{an APS `data.frame` as extracted by \code{eesectors::extract_APS_data}
and with sectors added with \code{eesectors::appendSectors} and
\code{eesectors:::sector_jobs}}
}
\value{
The function returns a `data.frame` contains the jobs per DCMS sector
}
\description{
Counts the number of jobs per sector, after removing those
  without a second or first jobs
}
\details{
Counts the number of jobs per sector, after removing those
  without a second or first jobs using the \code{eesectors:::main_job_mask}
  and \code{eesectors:::second_job_mask} functions.
}
\examples{

\dontrun{
library(eesectors)
job_counts_20xx=sector_jobs_count(APS_data_secotr_jobs)
)
}

}
