% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_sector_data.R
\name{year_sector_data}
\alias{year_sector_data}
\title{long data class}
\usage{
year_sector_data(x, log_level = futile.logger::WARN,
  log_appender = "console", log_issues = FALSE)
}
\arguments{
\item{x}{Input dataframe, see details.}

\item{log_level}{The severity level at which log messages are written from
least to most serious: TRACE, DEBUG, INFO, WARN, ERROR, FATAL. Default is
level is INFO. See \code{?flog.threshold()} for additional details.}

\item{log_appender}{Defaults to write the log to "console", alternatively you
can provide a character string to specify a filename to also write to. See
for additional details \code{?futile.logger::appender.tee()}.}

\item{log_issues}{should issues with the data quality be logged to github?
See \code{?raise_issue()} for additional details.}
}
\value{
If the class is instantiated correctly, nothing is returned.
}
\description{
\code{year_sector_data} is the class used for the creation of tables
  3.1, 4.1, 4.5, and 5.1 of the DCMS Sectors Economic Estimate
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf}).
}
\details{
The \code{year_sector_data} class expects a \code{data.frame} with three
  columns: sector, year, and measure, where measure is one of GVA, exports,
  or enterprises. The \code{data.frame} should include historical data, which
  is used for checks on the quality of this year\'s data, and for producing
  tables and plots.

  Once inititated, the class has five slots: \code{df}: the basic
  \code{data.frame}, \code{colnames}: a character vector containing the
  column names from the \code{df}, \code{type}: a character vector of
  \code{length(type) == 1} describing the type of data (\"GVA\", \"exports\",
  or \"enterprises\") imputed from column names, \code{sector_levels}: a
  factor vector containing levels of \code{df$sector} of the factor sector,
  \code{years}: an integer vector containing \code{unique(df$year)}.
}
\examples{

library(eesectors)

GVA <- year_sector_data(GVA_by_sector_2016)

}
