% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_sector_table.R
\name{year_sector_table}
\alias{year_sector_table}
\alias{year_sector_table.year_sector_data}
\title{year_sector_table()}
\usage{
year_sector_table(x, html, fmt, ...)

\method{year_sector_table}{year_sector_data}(x, html = FALSE, fmt = "\%.1f",
  ...)
}
\arguments{
\item{x}{Object of \code{class(x) == 'year_sector_data'}.}

\item{html}{Should the output be an R \code{data.frame} (and \code{tbl} and
\code{tbl_df}) or as html using \code{xtable}.}

\item{fmt}{Format for values in the table to be displayed as, following
\code{sprintf}.}

\item{...}{Passes arguments to \code{print.xtable} and \code{xtable}. Will
silently be dropped if \code{html = FALSE}.}
}
\value{
wide format table
}
\description{
Generic method to convert tables into wide format
}
\details{
Generic method to convert tables into wide format
}
\section{Methods (by class)}{
\itemize{
\item \code{year_sector_data}: Create wide table from year_sector_data() class
}}

\seealso{
\code{\link{sprintf}}
}
