% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{format_table()}
\usage{
format_table(x, html = TRUE)
}
\arguments{
\item{x}{Object of \code{class(x) == 'long_data'}.}

\item{html}{Should the output be an R \code{data.frame} (and \code{tbl} and
\code{tbl_df}) or as html using \code{xtable}.}
}
\value{
wide format table
}
\description{
Method to convert table into wide format
}
\details{
Method to convert table into wide format
}

