% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA_table.R
\name{GVA_table}
\alias{GVA_table}
\title{Format GVA table}
\usage{
GVA_table(x)
}
\arguments{
\item{base_size}{Integer. Sets the base size of text for the plot. Defaults
to \code{12}.}

\item{base_colour}{Character string. Sets the default colour of axes and axis
labels. Must be a named R colour or hexadecimal colour code (e.g.
"#FF0000"). Defaults to \code{gray40}.}

\item{axes}{Character string. Specifies the presence or absence of axes Must
be one of \code{n} (no axes), \code{x} (only x axis), \code{y} (only y
axis), or \code{xy} (both axes shown).}
}
\value{
Will not return anything of itself, but when used in conjuntion with
  \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the package
  \code{ggplot2}, will apply styling to a plot.
}
\description{
\code{GVA_table} Formats the GVA by sectors table labelled as
  3.1 in the 2016 Economic estimates of DCMS Sectors statistical release
  (\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/544103/DCMS_Sectors_Economic_Estimates_-_August_2016.pdf}).
}
\details{
\code{GVA_table} takes as input a standardised long format GVA data frame, and uses
  \code{xtable} to format a pretty html table.
}
\examples{

library(dplyr)
library(EESectors)

GVA_table(GVA)


}

