% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_to}
\alias{relative_to}
\title{Calculate relative values}
\usage{
relative_to(x, y, digits = 2)
}
\arguments{
\item{x}{base variable.}

\item{y}{relative variable.}

\item{digits}{Round output to this many digits.}
}
\value{
Returns the percentage difference of \code{y} relative to \code{x}.
}
\description{
Calculates the percentage difference in
  \code{y}, relative to \code{x}
}
\details{
Calculates the percentage difference in
  \code{y}, relative to \code{x}
}
